CuBoulder/ucb_site_configuration#36 reorganizes site settings (1)
