+++ ucb_site_configuration.links.menu.yml
@@ -18,8 +18,8 @@ ucb_site_configuration.appearance_form:
   parent: ucb_site_configuration
   weight: 1
 ucb_site_configuration.pages_form:
-  title: Pages and search
-  description: 'Change the site home page or search options.'
+  title: Content types
+  description: 'Change settings for specific content types.'
   route_name: ucb_site_configuration.pages_form
   parent: ucb_site_configuration
   weight: 2
@@ -29,15 +29,9 @@ ucb_site_configuration.contact_info_form:
   route_name: ucb_site_configuration.contact_info_form
   parent: ucb_site_configuration
   weight: 3
-ucb_site_configuration.related_articles_form:
-  title: Related articles
-  description: 'Exclude categories or tags from related articles.'
-  route_name: ucb_site_configuration.related_articles_form
-  parent: ucb_site_configuration
-  weight: 4
 ucb_site_configuration.external_services:
   title: Third-party services
   description: 'Configure third-party services to be used on the site.'
   route_name: entity.ucb_external_service_include.collection
   parent: ucb_site_configuration
-  weight: 5
+  weight: 4
