+++ ucb_site_configuration.links.task.yml
@@ -10,7 +10,7 @@ ucb_site_configuration.appearance_form:
   weight: 1
 ucb_site_configuration.pages_form:
   route_name: ucb_site_configuration.pages_form
-  title: 'Pages and search'
+  title: 'Content types'
   base_route: ucb_site_configuration
   weight: 2
 ucb_site_configuration.contact_info_form:
@@ -18,13 +18,8 @@ ucb_site_configuration.contact_info_form:
   title: 'Contact info'
   base_route: ucb_site_configuration
   weight: 3
-ucb_site_configuration.related_articles_form:
-  route_name: ucb_site_configuration.related_articles_form
-  title: 'Related articles'
-  base_route: ucb_site_configuration
-  weight: 4
 ucb_site_configuration.external_services:
   route_name: entity.ucb_external_service_include.collection
   title: 'Third-party services'
   base_route: ucb_site_configuration
-  weight: 5
+  weight: 4
